+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.2] - 2024-07-10
+### New
+- Added `SessionKeyValidator` module for ERC20 SessionKeyValidator.
+- Added `enableSessionKey`, `rotateSessionKey`, `disableSessionKey`, and `getAssociatedSessionKeys` functions to `SessionKeyValidator`.
 
 ## [1.1.1] - 2024-07-08
 ### Feature Enhancement